refactor: 온보딩 페이지 하나로 유지
