+++ src/entities/userInfo/index.ts
@@ -1,2 +1,3 @@
 export { getUserInfo } from './api/userApi'
 export { type UserInfo } from './api/types'
+export { useRefetchUserInfo } from './hooks/useRefetchUserInfo'
