+++ src/shared/lib/index.ts
@@ -2,3 +2,4 @@ export { useNavigater } from './hooks/useNavigater'
 export { useToast } from './hooks/useToast'
 export { getCookie } from './utils/cookies'
 export { getLocalStorage, setLocalStorage, removeLocalStorage } from './utils/localStorage'
+export { getSessionStorage, setSessionStorage, removeSessionStorage } from './utils/sessionStorage'
