+++ src/views/calendar/ui/AddEvent/AddEvent.tsx
@@ -168,7 +168,7 @@ export function AddEventPage() {
           ) : null
         }
       />
-      <div className="flex flex-col px-[1.5rem] overflow-hidden py-[1.5rem] h-[75vh] border-4">
+      <div className="flex flex-col px-[1.5rem] overflow-hidden py-[1.5rem] h-[75vh]">
         <div className="flex flex-col flex-grow overflow-y-auto gap-[3rem] ">
           <div className="relative flex justify-between items-center">
             <AddEventTitle placeholder={'Title'} />
